improvement(usepageviewtrigger): decouple page active check to usePageActiveTrigger
