+++ README.md
@@ -483,36 +483,74 @@ function Button({ onClick, 'tracking-label': label, children }) {
 
 ## PageView
 
-`usePageView(options)` lets you dispatch a page view event.
+### What can be considered a page view?
 
-The hooks accepts the following interface:
+- Page load
+- Route changes in SPAs
+- New "context" over the screen being displayed, such as modals.
 
-```jsx
-interface Options {
-  /**
-   * The trigger will automatically dispatch an event whenever the tab becomes
-   * inactive and then active again (via [Visibility change](https://developer.mozilla.org/en-US/docs/Web/API/Document/visibilitychange_event))
-   */
-  visibilityChange: boolean; // false by default
+The `pageView` event will be dispatched with:
+
+```ts
+interface Event {
+  app: string;
+  view: string;
+  event: 'page-view'; // Provided by available hooks
+  timestamp: number; // Provided by the library when the dispatch function is called
 }
 ```
 
+### Where to place the page view hook in your application
+
+In order to have a meaningful page view event, we recommend integrating the available hooks for page view after declaring the [TrackingRoot](#trackingroot) in your application.
+
+You don't need to declare it after the [TrackingView](#trackingview) since any `TrackingView` component will overwrite the context value.
+
+### Available hooks
+
+`usePageViewTrigger()` lets you dispatch a page view event.
+
 ```jsx
 import React from 'react';
-import { usePageViewTrigger } from '@sumup/collector';
+import {
+  TrackingRoot,
+  TrackingView,
+  usePageViewTrigger
+} from '@sumup/collector';
 
 interface Props {
   children: React.ReactNode;
   location: string;
 }
 
-function App({ location, children }: Props) {
-  const dispatchPageView = usePageViewTrigger({ visibilityChange: true });
+// This could be a hook instead
+function PageView({ location, children }: Props) {
+  const dispatchPageView = usePageViewTrigger();
 
   // run the effect everytime location changes
   useEffect(() => {
     dispatchPageView();
   }, [location]);
+
+  return children;
+}
+```
+
+`usePageActiveTrigger` **automatically** dispatchs an event whenever the tab becomes inactive and then active again (via [Visibility change](https://developer.mozilla.org/en-US/docs/Web/API/Document/visibilitychange_event)). This is meant to be used whenever you want to track if people are changing tabs.
+
+Keep in mind only one "pageActive" trigger is required since it's a document event listener.
+
+```jsx
+import React from 'react';
+import { usePageActiveTrigger } from '@sumup/collector';
+
+interface Props {
+  children: React.ReactNode;
+  location: string;
+}
+
+function PageActive({ location, children }: Props) {
+  usePageActiveTrigger();
 
   return children;
 }
