+++ src/hooks/usePageViewTrigger/usePageViewTrigger.spec.tsx
@@ -14,14 +14,14 @@
  */
 
 import * as React from 'react';
-import { render, fireEvent, act } from '@testing-library/react';
+import { render, fireEvent } from '@testing-library/react';
 
 import TrackingRoot from '../../components/TrackingRoot';
 import usePageViewTrigger from './usePageViewTrigger';
 import { Events } from '../../types';
 
 const DispatchButton = () => {
-  const dispatch = usePageViewTrigger({ visibilityChange: true });
+  const dispatch = usePageViewTrigger();
 
   return (
     <button data-testid="dispatch-btn" onClick={() => dispatch()}>
@@ -51,43 +51,5 @@ describe('usePageViewTrigger', () => {
     fireEvent.click(getByTestId(btn));
 
     expect(dispatch).toHaveBeenCalledWith(expected);
-  });
-  describe('when the document becomes inactive than active again', () => {
-    it('should provide a dispatch function that contains the pageView event', () => {
-      const dispatch = jest.fn();
-      const app = 'test-app-hook';
-
-      const expected = {
-        app,
-        event: Events.pageView,
-        timestamp: expect.any(Number)
-      };
-
-      render(
-        <TrackingRoot name={app} onDispatch={dispatch}>
-          <DispatchButton />
-        </TrackingRoot>
-      );
-
-      act(() => {
-        Object.defineProperty(document, 'hidden', {
-          configurable: true,
-          value: true
-        });
-
-        document.dispatchEvent(new Event('visibilitychange'));
-      });
-
-      act(() => {
-        Object.defineProperty(document, 'hidden', {
-          configurable: true,
-          value: false
-        });
-
-        document.dispatchEvent(new Event('visibilitychange'));
-      });
-
-      expect(dispatch).toHaveBeenCalledWith(expected);
-    });
   });
 });
